# backend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: Good Doggy CRM API
  version: 0.1.0
  description: Contrato de la API del backend. Fuente para el cliente TypeScript (orval).
paths:
  /api/v1/auth/csrf/:
    get:
      operationId: auth_csrf
      description: Entrega la cookie `csrftoken` antes del primer método no seguro.
      tags:
      - auth
      responses:
        '204':
          description: No response body
  /api/v1/auth/login/:
    post:
      operationId: auth_login
      description: Inicia sesión. `INVALID_CREDENTIALS` no distingue el motivo del
        rechazo.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/v1/auth/logout/:
    post:
      operationId: auth_logout
      tags:
      - auth
      security:
      - sessionCookie: []
      responses:
        '204':
          description: No response body
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/v1/auth/session/:
    get:
      operationId: auth_session
      description: |-
        La sesión actual. 401 `NOT_AUTHENTICATED` si no hay, caducó o su usuario no está
        activo.
      tags:
      - auth
      security:
      - sessionCookie: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/v1/me/organizations/:
    get:
      operationId: me_organizations
      description: |-
        Organizaciones del usuario: membresía activa y organización no suspendida. No abre
        un `tenant_scope`: lee con `user_scope`, que solo deja ver las membresías propias.
      tags:
      - me
      security:
      - sessionCookie: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationSummary'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
components:
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          description: Código estable. Lo único que decide el cliente.
        message:
          type: string
          description: Texto seguro para mostrar.
        fields:
          type: object
          additionalProperties: {}
          description: 'Solo en VALIDATION_ERROR: errores por campo (`_`: generales).'
      required:
      - code
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
          maxLength: 254
        password:
          type: string
          writeOnly: true
          minLength: 1
          maxLength: 1024
      required:
      - email
      - password
    OrganizationSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
      required:
      - id
      - name
      - slug
    Session:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
      - user
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        first_name:
          type: string
          readOnly: true
        last_name:
          type: string
          readOnly: true
      required:
      - email
      - first_name
      - id
      - last_name
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: crm_session
      description: Cookie de sesión (`__Host-crm_session` en producción). Los métodos
        no seguros llevan además la cabecera `X-CSRFToken`.